requirements.txt
langchain
langchain_openai
langchain_community
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
tiktoken